Utilities and data structures

Active set

FrankWolfe.ActiveSet — Type

ActiveSet{AT, R, IT}

Represents an active set of extreme vertices collected in a FW algorithm, along with their coefficients (λ_i, a_i). R is the type of the λ_i, AT is the type of the atoms a_i. The iterate x = ∑λ_i a_i is stored in x with type IT.

Base.copy — Method

Copies an active set, the weight and atom vectors and the iterate. Individual atoms are not copied.

FrankWolfe.active_set_argmin — Method

active_set_argmin(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_i, a_i, i)

FrankWolfe.active_set_argminmax — Method

active_set_argminmax(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_min, a_min, i_min, val_min, λ_max, a_max, i_max, val_max, val_max-val_min ≥ Φ)

FrankWolfe.active_set_initialize! — Method

active_set_initialize!(as, v)

Resets the active set structure to a single vertex v with unit weight.

FrankWolfe.active_set_update! — Function

active_set_update!(active_set::ActiveSet, lambda, atom)

Adds the atom to the active set with weight lambda or adds lambda to existing atom.

FrankWolfe.active_set_update_iterate_pairwise! — Method

active_set_update_iterate_pairwise!(x, lambda, fw_atom, away_atom)

Operates x ← x + λ a_fw - λ a_aw.

FrankWolfe.active_set_update_scale! — Method

active_set_update_scale!(x, lambda, atom)

Operates x ← (1-λ) x + λ a.

FrankWolfe.get_active_set_iterate — Method

get_active_set_iterate(active_set)

Return the current iterate corresponding. Does not recompute it.

Functions and gradients

FrankWolfe.ObjectiveFunction — Type

ObjectiveFunction

Represents an objective function optimized by algorithms. Subtypes of ObjectiveFunction must implement at least

compute_value(::ObjectiveFunction, x) for primal value evaluation

compute_gradient(::ObjectiveFunction, x) for gradient evaluation.

and optionally compute_value_gradient(::ObjectiveFunction, x) returning the (primal, gradient) pair. compute_gradient may always use the same storage and return a reference to it.

FrankWolfe.SimpleFunctionObjective — Type

SimpleFunctionObjective{F,G,S}

An objective function built from separate primal objective f(x) and in-place gradient function grad!(storage, x). It keeps an internal storage of type s used to evaluate the gradient in-place.

FrankWolfe.StochasticObjective — Type

StochasticObjective{F, G, XT, S}(f::F, grad!::G, xs::XT, storage::S)

Represents a composite function evaluated with stochastic gradient. f(θ, x) evaluates the loss for a single data point x and parameter θ. grad!(storage, θ, x) adds to storage the partial gradient with respect to data point x at parameter θ. xs must be an indexable iterable (Vector{Vector{Float64}} for instance). Functions using a StochasticObjective have optional keyword arguments rng, batch_size and full_evaluation controlling whether the function should be evaluated over all data points.

Note: grad! must not reset the storage to 0 before adding to it.

FrankWolfe.compute_gradient — Function

compute_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes the gradient of f at x. May return a reference to an internal storage.

FrankWolfe.compute_value — Function

compute_value(f::ObjectiveFunction, x; [kwargs...])

Computes the objective f at x.

FrankWolfe.compute_value_gradient — Method

compute_value_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes in one call the pair (value, gradient) evaluated at x. By default, calls compute_value and compute_gradient with keywords kwargs passed down to both.

Callbacks

FrankWolfe.CallbackState — Type

Main structure created before and passed to the callback in first position.

Custom vertex storage

Custom extreme point types

For some feasible sets, the extreme points of the feasible set returned by the LMO possess a specific structure that can be represented in an efficient manner both for storage and for common operations like scaling and addition with an iterate. They are presented below:

FrankWolfe.ScaledHotVector — Type

ScaledHotVector{T}

Represents a vector of at most one value different from 0.

FrankWolfe.RankOneMatrix — Type

RankOneMatrix{T, UT, VT}

Represents a rank-one matrix R = u * vt'. Composes like a charm.

Utils

FrankWolfe.ConstantBatchIterator — Type

ConstantBatchIterator(batch_size)

Batch iterator always returning a constant batch size.

FrankWolfe.ConstantMomentumIterator — Type

ConstantMomentumIterator{T}

Iterator for momentum with a fixed damping value, always return the value and a dummy state.

FrankWolfe.DeletedVertexStorage — Type

Vertex storage to store dropped vertices or find a suitable direction in lazy settings. The algorithm will look for at most return_kth suitable atoms before returning the best. See Extra-lazification with a vertex storage for usage.

A vertex storage can be any type that implements two operations:

Base.push!(storage, atom) to add an atom to the storage.

Note that it is the storage type responsibility to ensure uniqueness of the atoms present.

storage_find_argmin_vertex(storage, direction, lazy_threshold) -> (found, vertex)

returning whether a vertex with sufficient progress was found and the vertex. It is up to the storage to remove vertices (or not) when they have been picked up.

FrankWolfe.ExpMomentumIterator — Type

ExpMomentumIterator{T}

Iterator for the momentum used in the variant of Stochastic Frank-Wolfe. Momentum coefficients are the values of the iterator: ρ_t = 1 - num / (offset + t)^exp

The state corresponds to the iteration count.

Source: Stochastic Conditional Gradient Methods: From Convex Minimization to Submodular Maximization Aryan Mokhtari, Hamed Hassani, Amin Karbasi, JMLR 2020.

FrankWolfe.IncrementBatchIterator — Type

IncrementBatchIterator(starting_batch_size, max_batch_size, [increment = 1])

Batch size starting at startingbatchsize and incrementing by increment at every iteration.

FrankWolfe._unsafe_equal — Method

_unsafe_equal(a, b)

Like isequal on arrays but without the checks. Assumes a and b have the same axes.

FrankWolfe.batchsize_iterate — Function

batchsize_iterate(iter::BatchSizeIterator) -> b

Method to implement for a batch size iterator of type BatchSizeIterator. Calling batchsize_iterate returns the next batch size and typically update the internal state of iter.

FrankWolfe.momentum_iterate — Function

momentum_iterate(iter::MomentumIterator) -> ρ

Method to implement for a type MomentumIterator. Returns the next momentum value ρ and updates the iterator internal state.

FrankWolfe.muladd_memory_mode — Method

muladd_memory_mode(memory_mode::MemoryEmphasis, d, x, v)

Performs d = x - v in-place or not depending on MemoryEmphasis

FrankWolfe.muladd_memory_mode — Method

(memory_mode::MemoryEmphasis, storage, x, gamma::Real, d)

Performs storage = x - gamma * d in-place or not depending on MemoryEmphasis

FrankWolfe.muladd_memory_mode — Method

(memory_mode::MemoryEmphasis, x, gamma::Real, d)

Performs x = x - gamma * d in-place or not depending on MemoryEmphasis

FrankWolfe.storage_find_argmin_vertex — Method

Give the vertex v in the storage that minimizes s = direction ⋅ v and whether s achieves s ≤ lazy_threshold.

FrankWolfe.trajectory_callback — Method

trajectory_callback(storage)

Callback pushing the state at each iteration to the passed storage. The state data is only the 5 first fields, usually: (t,primal,dual,dual_gap,time)

Oracle counting trackers

The following structures are wrapping given oracles to behave similarly but additionally track the number of calls.

FrankWolfe.TrackingObjective — Type

A function acting like the normal objective f but tracking the number of calls.

FrankWolfe.TrackingGradient — Type

A function acting like the normal grad! but tracking the number of calls.

FrankWolfe.TrackingLMO — Type

TrackingLMO{LMO}(lmo)

An LMO wrapping another one and tracking the number of calls.

Also see the example "Tracking number of calls to different oracles".

Index

FrankWolfe.ActiveSet

FrankWolfe.CallbackState

FrankWolfe.ConstantBatchIterator

FrankWolfe.ConstantMomentumIterator

FrankWolfe.DeletedVertexStorage

FrankWolfe.ExpMomentumIterator

FrankWolfe.IncrementBatchIterator

FrankWolfe.ObjectiveFunction

FrankWolfe.RankOneMatrix

FrankWolfe.ScaledHotVector

FrankWolfe.SimpleFunctionObjective

FrankWolfe.StochasticObjective

FrankWolfe.TrackingGradient

FrankWolfe.TrackingLMO

FrankWolfe.TrackingObjective

Base.copy

FrankWolfe._unsafe_equal

FrankWolfe.active_set_argmin

FrankWolfe.active_set_argminmax

FrankWolfe.active_set_initialize!

FrankWolfe.active_set_update!

FrankWolfe.active_set_update_iterate_pairwise!

FrankWolfe.active_set_update_scale!

FrankWolfe.batchsize_iterate

FrankWolfe.compute_gradient

FrankWolfe.compute_value

FrankWolfe.compute_value_gradient

FrankWolfe.get_active_set_iterate

FrankWolfe.momentum_iterate

FrankWolfe.muladd_memory_mode

FrankWolfe.muladd_memory_mode

FrankWolfe.muladd_memory_mode

FrankWolfe.storage_find_argmin_vertex

FrankWolfe.trajectory_callback