Algorithms

This section contains all algorithms of the FrankWolfe.jl package.

Functions

Missing docstring.

Missing docstring for blended_conditional_gradient. Check Documenter's build log for details.

FrankWolfe.compute_extreme_point — Function

compute_extreme_point(lmo::LinearMinimizationOracle, direction; kwargs...)

Computes the point argmin_{v ∈ C} v ⋅ direction with C the set represented by the LMO. All LMOs should accept keyword arguments that they can ignore.

Missing docstring.

Missing docstring for frank_wolfe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for lazified_conditional_gradient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for away_frank_wolfe. Check Documenter's build log for details.

LMOs

The Linear Minimization Oracle (LMO) is an integral part of the iterative step in the FW algorithm. Given $d\in \mathcal{X}$, it returns:

\[v\in \argmin_{x\in \mathcal{C}} \langle d,x \rangle.\]

FrankWolfe.jl features the following common LMOs out of the box:

probability simplex: FrankWolfe.ProbabilitySimplexOracle

unit simplex: FrankWolfe.UnitSimplexOracle

$K$-sparse polytope: FrankWolfe.KSparseLMO

$K$-norm ball: FrankWolfe.KNormBallLMO

$L^p$-norm ball: FrankWolfe.LpNormLMO

Birkhoff polytope: FrankWolfe.BirkhoffPolytopeLMO

All of them are subtypes of FrankWolfe.LinearMinimizationOracle and implement the compute_extreme_point method.

Functions and Structures

FrankWolfe.LinearMinimizationOracle — Type

Supertype for linear minimization oracles.

All LMOs must implement compute_extreme_point(lmo::LMO, direction) and return a vector v of the appropriate type.

FrankWolfe.CachedLinearMinimizationOracle — Type

CachedLinearMinimizationOracle{LMO}

Oracle wrapping another one of type lmo. Subtypes of CachedLinearMinimizationOracle contain a cache of previous solutions.

By convention, the inner oracle is named inner. Cached optimizers are expected to implement Base.empty! and Base.length.

FrankWolfe.SingleLastCachedLMO — Type

SingleLastCachedLMO{LMO, VT}

Caches only the last result from an LMO and stores it in last_vertex. Vertices of LMO have to be of type VT if provided.

FrankWolfe.MultiCacheLMO — Type

MultiCacheLMO{N, LMO, VT}

Cache for a LMO storing up to N vertices in the cache, removed in FIFO style. oldest_idx keeps track of the oldest index in the tuple, i.e. to replace next. VT, if provided, must be the type of vertices returned by LMO

FrankWolfe.VectorCacheLMO — Type

VectorCacheLMO{N, LMO, VT}

Cache for a LMO storing an unbounded number of vertices of type VT in the cache. VT, if provided, must be the type of vertices returned by LMO

FrankWolfe.ProductLMO — Type

ProductLMO(lmos...)

Linear minimization oracle over the Cartesian product of multiple LMOs.

FrankWolfe.compute_extreme_point — Method

compute_extreme_point(lmo::ProductLMO, direction::Tuple; kwargs...)

Extreme point computation on Cartesian product, with a direction (d1, d2, ...) given as a tuple of directions. All keyword arguments are passed to all LMOs.

FrankWolfe.compute_extreme_point — Method

compute_extreme_point(lmo::ProductLMO, direction::AbstractArray; direction_indices, storage=similar(direction))

Extreme point computation, with a direction array and direction_indices provided such that: direction[direction_indices[i]] is passed to the i-th LMO. The result is stored in the optional storage container.

All keyword arguments are passed to all LMOs.

FrankWolfe.UnitSimplexOracle — Type

UnitSimplexOracle(right_side)

Represents the scaled unit simplex:

C = {x ∈ R^n_+, ∑x ≤ right_side}

FrankWolfe.compute_extreme_point — Method

LMO for scaled unit simplex: ∑ x_i = τ Returns either vector of zeros or vector with one active value equal to RHS if there exists an improving direction.

FrankWolfe.compute_dual_solution — Method

Dual costs for a given primal solution to form a primal dual pair for scaled unit simplex. Returns two vectors. The first one is the dual costs associated with the constraints and the second is the reduced costs for the variables.

FrankWolfe.ProbabilitySimplexOracle — Type

ProbabilitySimplexOracle(right_side)

Represents the scaled probability simplex:

C = {x ∈ R^n_+, ∑x = right_side}

FrankWolfe.compute_extreme_point — Method

LMO for scaled probability simplex. Returns a vector with one active value equal to RHS in the most improving (or least degrading) direction.

FrankWolfe.compute_dual_solution — Method

Dual costs for a given primal solution to form a primal dual pair for scaled probability simplex. Returns two vectors. The first one is the dual costs associated with the constraints and the second is the reduced costs for the variables.

FrankWolfe.KSparseLMO — Type

KSparseLMO{T}(K::Int, right_hand_side::T)

LMO for the K-sparse polytope:

C = B_1(τK) ∩ B_∞(τ)

with τ the right_hand_side parameter. The LMO results in a vector with the K largest absolute values of direction, taking values -τ sign(x_i).

FrankWolfe.BirkhoffPolytopeLMO — Type

BirkhoffPolytopeLMO

The Birkhoff polytope encodes doubly stochastic matrices. Its extreme vertices are all permutation matrices of side-dimension dimension.

FrankWolfe.LpNormLMO — Type

LpNormLMO{T, p}(right_hand_side)

LMO with feasible set being a bound on the L-p norm:

C = {x ∈ R^n, norm(x, p) ≤ right_side}

FrankWolfe.KNormBallLMO — Type

KNormBallLMO{T}(K::Int, right_hand_side::T)

LMO for the K-norm ball, intersection of L1-ball (τK) and L∞-ball (τ/K)

C_{K,τ} = conv { B_1(τ) ∪ B_∞(τ / K) }

with τ the right_hand_side parameter.

FrankWolfe.NuclearNormLMO — Type

NuclearNormLMO{T}(δ)

LMO over matrices that have a nuclear norm less than δ. The LMO returns the rank-one matrix with singular value δ.

FrankWolfe.compute_extreme_point — Method

Best rank-one approximation using the Golub-Kahan-Lanczos bidiagonalization from IterativeSolvers.

Warning: this does not work (yet) with all number types, BigFloat and Float16 fail.

FrankWolfe.MathOptLMO — Type

MathOptLMO{OT <: MOI.Optimizer} <: LinearMinimizationOracle

Linear minimization oracle with feasible space defined through a MathOptInterface.Optimizer. The oracle call sets the direction and reruns the optimizer.

The direction vector has to be set in the same order of variables as the MOI.ListOfVariableIndices() getter.

FrankWolfe.convert_mathopt — Function

convert_mathopt(lmo::LMO, optimizer::OT; kwargs...) -> MathOptLMO{OT}

Converts the given LMO to its equivalent MathOptInterface representation using optimizer. Must be implemented by LMOs.

Components

This section gathers all additional relevant components of the FrankWolfe.jl package.

Functions and Structures

FrankWolfe.ActiveSet — Type

ActiveSet{AT, R, IT}

Represents an active set of extreme vertices collected in a FW algorithm, along with their coefficients (λ_i, a_i). R is the type of the λ_i, AT is the type of the atoms a_i. The iterate x = ∑λ_i a_i is stored in x with type IT.

FrankWolfe.active_set_update! — Function

active_set_update!(active_set::ActiveSet, lambda, atom)

Adds the atom to the active set with weight lambda or adds lambda to existing atom.

FrankWolfe.compute_active_set_iterate — Function

compute_active_set_iterate(active_set)

FrankWolfe.active_set_argmin — Function

active_set_argmin(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_i, a_i, i)

FrankWolfe.active_set_argminmax — Function

active_set_argminmax(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_min, a_min, i_min, λ_max, a_max, i_max)

FrankWolfe.find_minmax_directions — Function

find_minmax_directions(active_set::ActiveSet, direction, Φ)

Computes the point of the active set minimizing in direction on the active set (local Frank Wolfe) and the maximizing one (away step). Returns the two corresponding indices in the active set, along with a flag indicating if the direction improvement is above a threshold. goodstep_tolerance ∈ (0, 1] is a tolerance coefficient multiplying Φ for the validation of the progress.

FrankWolfe.minimize_over_convex_hull! — Function

minimize_over_convex_hull!

Given a function f with gradient grad! and an active set active_set this function will minimize the function over the convex hull of the active set until the strong-wolfe gap over the active set is below tolerance.

It will either directly minimize over the convex hull using simplex gradient descent, or it will transform the problem to barycentric coordinates and minimize over the unit probability simplex using gradient descent or Nesterov's accelerated gradient descent.

FrankWolfe.build_reduced_problem — Method

build_reduced_problem(atoms::AbstractVector{<:FrankWolfe.ScaledHotVector}, hessian, weights, gradient, tolerance)

Given an active set formed by ScaledHotVector, a (constant) Hessian and a gradient constructs a quadratic problem over the unit probability simplex that is equivalent to minimizing the original function over the convex hull of the active set. If λ are the barycentric coordinates of dimension equal to the cardinality of the active set, the objective function is: f(λ) = reducedlinear^T λ + 0.5 * λ^T reducedhessian λ

In the case where we find that the current iterate has a strong-Wolfe gap over the convex hull of the active set that is below the tolerance we return nothing (as there is nothing to do).

FrankWolfe.build_reduced_problem — Method

buildreducedproblem

Same as the function above, but for the case where the active set is formed by Sparse Arrays.

FrankWolfe.build_reduced_problem — Method

build_reduced_problem(atoms::AbstractVector{<:Array}, hessian, weights, gradient, tolerance)

Same as the two function above, but for the case where the active set is formed by dense Arrays.

FrankWolfe.Strong_Frank_Wolfe_gap — Function

Checks the strong Frank-Wolfe gap for the reduced problem.

FrankWolfe.accelerated_simplex_gradient_descent_over_probability_simplex — Function

accelerated_simplex_gradient_descent_over_probability_simplex

Minimizes an objective function over the unit probability simplex until the Strong-Wolfe gap is below tolerance using Nesterov's accelerated gradient descent.

FrankWolfe.simplex_gradient_descent_over_probability_simplex — Function

simplexgradientdescentoverprobability_simplex

Minimizes an objective function over the unit probability simplex until the Strong-Wolfe gap is below tolerance using gradient descent.

FrankWolfe.projection_simplex_sort — Function

projection_simplex_sort(x; s=1.0)

Perform a projection onto the probability simplex of radius s using a sorting algorithm.

FrankWolfe.Strong_Frank_Wolfe_gap_probability_simplex — Function

StrongFrankWolfegapprobability_simplex

Compute the Strong-Wolfe gap over the unit probability simplex given a gradient.

FrankWolfe.simplex_gradient_descent_over_convex_hull — Function

simplexgradientdescentoverconvex_hull

Minimizes an objective function over the convex hull of the active set until the Strong-Wolfe gap is below tolerance using simplex gradient descent.

FrankWolfe.lp_separation_oracle — Function

Returns either a tuple (y, val) with y an atom from the active set satisfying the progress criterion and val the corresponding gap dot(y, direction) or the same tuple with y from the LMO.

inplace_loop controls whether the iterate type allows in-place writes. kwargs are passed on to the LMO oracle.

FrankWolfe.LineSearchMethod — Type

Line search method to apply once the direction is computed.

FrankWolfe.Emphasis — Type

Emphasis given to the algorithm for memory-saving or not. The memory-saving mode may not be faster than the default blas mode for small dimensions.

FrankWolfe.ObjectiveFunction — Type

ObjectiveFunction{VT}

Represents an objective function optimized by algorithms. Subtypes of ObjectiveFunction must implement at least

compute_value(::ObjectiveFunction, x) for primal value evaluation

compute_gradient(::ObjectiveFunction, x) for gradient evaluation.

and optionally compute_value_gradient(::ObjectiveFunction, x) returning the (primal, gradient) pair.

FrankWolfe.compute_value_gradient — Function

compute_value_gradient(f::ObjectiveFunction, x; kwargs)

Computes in one call the pair (function_value, function_grad) evaluated at x. By default, calls compute_value and compute_gradient with keyword kwargs passed to both.

FrankWolfe.StochasticObjective — Type

StochasticObjective{F, G, XT}(f::F, grad::G, xs::XT)

Represents an objective function evaluated with stochastic gradient. f(θ, x) evaluates the loss for data point x and parameter θ. grad(θ, x) evaluates the loss gradient with respect to data point x at parameter θ. xs must be an indexable iterable (Vector{Vector{Float64}} for instance). Functions using a StochasticObjective have optional keyword arguments rng, batch_size and full_evaluation controlling whether the function should be evaluated over all data points.

FrankWolfe.ScaledHotVector — Type

ScaledHotVector{T}

Represents a vector of at most one value different from 0.

FrankWolfe.RankOneMatrix — Type

RankOneMatrix{T, UT, VT}

Represents a rank-one matrix R = u * vt'. Composes like a charm.

FrankWolfe.line_search_wrapper — Function

line search wrapper NOTE: The stepsize is defined as x - gamma * d

Returns the step size gamma and the Lipschitz estimate L

FrankWolfe.adaptive_step_size — Function

Slight modification of Adaptive Step Size strategy from https://arxiv.org/pdf/1806.05123.pdf

Note: direction is opposite to the improving direction norm(gradient, direction) > 0 TODO:

make emphasis aware and optimize

FrankWolfe.plot_results — Function

plot_results

Given a series of list, generate subplots. listdatay -> contains a list of a list of lists (where each list refers to a subplot, and a list of lists refers to the y-values of the series inside a subplot). listdatax -> contains a list of a list of lists (where each list refers to a subplot, and a list of lists refers to the x-values of the series inside a subplot). So if we have one plot with two series, these might look like: listdatay = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]]] listdatax = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]]]

And if we have two plots, each with two series, these might look like: listdatay = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]], [[7, 8, 9, 10, 11, 12], [7, 8, 9, 10, 11, 12]]] listdatax = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]], [[7, 8, 9, 10, 11, 12], [7, 8, 9, 10, 11, 12]]]

listlabel -> contains the labels for the series that will be plotted, which has to have a length equal to the number of series that are being plotted: listlabel = ["Series 1", "Series 2"]

listaxisx -> contains the labels for the x-axis that will be plotted, which has to have a length equal to the number of subplots: listaxisx = ["x-axis plot 1", "x-axis plot 1"]

listaxisy -> Same as listaxisx but for the y-axis

xscalelog -> A list of values indicating the type of axes to use in each subplot, must be equal to the number of subplots: xscalelog = [:log, :identity]

yscalelog -> Same as xscalelog but for the y-axis

FrankWolfe._unsafe_equal — Function

isequal without the checks. Assumes a and b have the same axes.

FrankWolfe.check_gradients — Function

Check if the gradient using finite differences matches the grad! provided.

FrankWolfe.trajectory_callback — Function

trajectory_callback(storage)

Callback pushing the state at each iteration to the passed storage. The state data is only the 5 first fields, usually: (t,primal,dual,dual_gap,time)