Line search and step size settings

The step size dictates how far one traverses along a local descent direction. More specifically, the step size $gamma_t$ is used at each iteration to determine how much the next iterate moves towards the new vertex:

\[x_{t+1} = x_t - \gamma_t (x_t - v_t).\]

$\gamma_t = 1$ implies that the next iterate is exactly the vertex, a zero $\gamma_t$ implies that the iterate is not moving.

The following are step size selection rules for Frank Wolfe algorithms. Some methodologies (e.g. FixedStep and Agnostic) depend only on the iteration number and induce series $\gamma_t$ that are independent of the problem data, while others (e.g. GoldenSearch and Adaptive) change according to local information about the function; the adaptive methods often require extra function and/or gradient computations. The typical options for convex optimization are Agnostic or Adaptive.

All step size computation strategies are subtypes of FrankWolfe.LineSearchMethod. The key method they have to implement is FrankWolfe.perform_line_search which is called at every iteration to compute the step size gamma.

FrankWolfe.LineSearchMethod — Type

Line search method to apply once the direction is computed. A LineSearchMethod must implement

perform_line_search(ls::LineSearchMethod, t, f, grad!, gradient, x, d, gamma_max, workspace)

with d = x - v. It may also implement build_linesearch_workspace(x, gradient) which creates a workspace structure that is passed as last argument to perform_line_search.

FrankWolfe.perform_line_search — Function

perform_line_search(ls::LineSearchMethod, t, f, grad!, gradient, x, d, gamma_max, workspace)

Returns the step size gamma for step size strategy ls.

FrankWolfe.Adaptive — Type

Slight modification of the Adaptive Step Size strategy from this paper

The Adaptive struct keeps track of the Lipschitz constant estimate L_est. perform_line_search also has a should_upgrade keyword argument on whether there should be a temporary upgrade to BigFloat for extended precision.

FrankWolfe.Agnostic — Type

Computes step size: 2/(2 + t) at iteration t.

FrankWolfe.Backtracking — Type

Backtracking(limit_num_steps, tol, tau)

Backtracking line search strategy, see this reference.

FrankWolfe.FixedStep — Type

Fixed step size strategy. The step size can still be truncated by the gamma_max argument.

FrankWolfe.Goldenratio — Type

Goldenratio

Simple golden-ratio based line search Golden Section Search, based on the Boosted FW paper code and adapted.

FrankWolfe.MonotonicNonConvexStepSize — Type

MonotonicNonConvexStepSize{F}

Represents a monotonic open-loop non-convex step size. Contains a halving factor N increased at each iteration until there is primal progress gamma = 1 / sqrt(t + 1) * 2^(-N).

FrankWolfe.MonotonicStepSize — Type

MonotonicStepSize{F}

Represents a monotonic open-loop step size. Contains a halving factor N increased at each iteration until there is primal progress gamma = 2 / (t + 2) * 2^(-N).

FrankWolfe.Nonconvex — Type

Computes a step size for nonconvex functions: 1/sqrt(t + 1).

FrankWolfe.Shortstep — Type

Computes the 'Short step' step size: dual_gap / (L * norm(x - v)^2), where L is the Lipschitz constant of the gradient, x is the current iterate, and v is the current Frank-Wolfe vertex.

See Pedregosa, Negiar, Askari, Jaggi 2020 for the adaptive step size, Carderera, Besançon, Pokutta 2021 for the monotonic step size.

Index

FrankWolfe.Adaptive

FrankWolfe.Agnostic

FrankWolfe.Backtracking

FrankWolfe.FixedStep

FrankWolfe.Goldenratio

FrankWolfe.LineSearchMethod

FrankWolfe.MonotonicNonConvexStepSize

FrankWolfe.MonotonicStepSize

FrankWolfe.Nonconvex

FrankWolfe.Shortstep

FrankWolfe.perform_line_search