Intersection problems

Problems over intersections of convex sets, i.e.

\[\min_{x \in \bigcap_{i=1}^n P_i} f(x),\]

pose a challenge as one has to combine the information of two or more LMOs. The key idea of AlternatingLinearMinimization is to convert the problem into a series of subproblems over single sets. To find a point within the intersection, one minimizes both the distance to the iterates of the other subproblems and the original objective function.

FrankWolfe.AlternatingLinearMinimization calls a block-coordinate method to solve the subproblems parallelly. All block-coordinate methods are subtypes of FrankWolfe.BlockCoordinateMethod and implement FrankWolfe.perform_bc_updates which solves the subproblems.

Missing docstring.

Missing docstring for FrankWolfe.AlternatingLinearMinimization. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FrankWolfe.BlockCoordinateMethod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FrankWolfe.perform_bc_updates. Check Documenter's build log for details.

FrankWolfe.alternating_linear_minimization — Method

alm(bc_algo, f, grad!, lmos, x0; ...)

Alternating Linear Minimizations Frank-Wolfe algorithm. Returns a tuple (x, v, primal, dual_gap, infeas, traj_data) with:

x cartesian product of final iterates

v cartesian product of last vertices of the LMOs

primal primal value f(x)

dual_gap final Frank-Wolfe gap

infeas sum of squared, pairwise distances between iterates

traj_data vector of trajectory information.

See Beck, Pauwels, Sabach 2015 for more details about the underlying theory.

Index

FrankWolfe.alternating_linear_minimization