Utilities and data structures

Active set

Functions and gradients

FrankWolfe.SimpleFunctionObjective — Type

SimpleFunctionObjective{F,G,S}

An objective function built from separate primal objective f(x) and in-place gradient function grad!(storage, x). It keeps an internal storage of type s used to evaluate the gradient in-place.

FrankWolfe.compute_gradient — Function

compute_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes the gradient of f at x. May return a reference to an internal storage.

FrankWolfe.compute_value — Function

compute_value(f::ObjectiveFunction, x; [kwargs...])

Computes the objective f at x.

FrankWolfe.compute_value_gradient — Method

compute_value_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes in one call the pair (value, gradient) evaluated at x. By default, calls compute_value and compute_gradient with keywords kwargs passed down to both.

Custom vertex storage

Utils

FrankWolfe._unsafe_equal — Method

_unsafe_equal(a, b)

Like isequal on arrays but without the checks. Assumes a and b have the same axes.

FrankWolfe.muladd_memory_mode — Method

muladd_memory_mode(memory_mode::MemoryEmphasis, d, x, v)

Performs d = x - v in-place or not depending on MemoryEmphasis

FrankWolfe.muladd_memory_mode — Method

(memory_mode::MemoryEmphasis, storage, x, gamma::Real, d)

Performs storage = x - gamma * d in-place or not depending on MemoryEmphasis

FrankWolfe.muladd_memory_mode — Method

(memory_mode::MemoryEmphasis, x, gamma::Real, d)

Performs x = x - gamma * d in-place or not depending on MemoryEmphasis

FrankWolfe.trajectory_callback — Method

trajectory_callback(storage)

Callback pushing the state at each iteration to the passed storage. The state data is only the 5 first fields, usually: (t,primal,dual,dual_gap,time)

Index

FrankWolfe.SimpleFunctionObjective

FrankWolfe._unsafe_equal

FrankWolfe.compute_gradient

FrankWolfe.compute_value

FrankWolfe.compute_value_gradient

FrankWolfe.muladd_memory_mode

FrankWolfe.muladd_memory_mode

FrankWolfe.muladd_memory_mode

FrankWolfe.trajectory_callback