Utilities and data structures

Active set

FrankWolfe.ActiveSet — Type

ActiveSet{AT, R, IT}

Represents an active set of extreme vertices collected in a FW algorithm, along with their coefficients (λ_i, a_i). R is the type of the λ_i, AT is the type of the atoms a_i. The iterate x = ∑λ_i a_i is stored in x with type IT.

Base.copy — Method

Copies an active set, the weight and atom vectors and the iterate. Individual atoms are not copied.

FrankWolfe.active_set_argmin — Method

active_set_argmin(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_i, a_i, i)

FrankWolfe.active_set_argminmax — Method

active_set_argminmax(active_set::ActiveSet, direction)

Computes the linear minimizer in the direction on the active set. Returns (λ_min, a_min, i_min, λ_max, a_max, i_max)

FrankWolfe.active_set_initialize! — Method

active_set_initialize!(as, v)

Resets the active set structure to a single vertex v with unit weight.

FrankWolfe.active_set_update! — Function

active_set_update!(active_set::ActiveSet, lambda, atom)

Adds the atom to the active set with weight lambda or adds lambda to existing atom.

FrankWolfe.compute_active_set_iterate — Method

compute_active_set_iterate(active_set)

FrankWolfe.find_minmax_directions — Method

find_minmax_directions(active_set::ActiveSet, direction, Φ)

Computes the point of the active set minimizing in direction on the active set (local Frank Wolfe) and the maximizing one (away step). Returns the two corresponding indices in the active set, along with a flag indicating if the direction improvement is above a threshold. goodstep_tolerance ∈ (0, 1] is a tolerance coefficient multiplying Φ for the validation of the progress.

Functions and gradients

FrankWolfe.ObjectiveFunction — Type

ObjectiveFunction

Represents an objective function optimized by algorithms. Subtypes of ObjectiveFunction must implement at least

compute_value(::ObjectiveFunction, x) for primal value evaluation

compute_gradient(::ObjectiveFunction, x) for gradient evaluation.

and optionally compute_value_gradient(::ObjectiveFunction, x) returning the (primal, gradient) pair. compute_gradient may always use the same storage and return a reference to it.

FrankWolfe.SimpleFunctionObjective — Type

SimpleFunctionObjective{F,G,S}

An objective function built from separate primal objective f(x) and in-place gradient function grad!(storage, x). It keeps an internal storage of type s used to evaluate the gradient in-place.

FrankWolfe.StochasticObjective — Type

StochasticObjective{F, G, XT, S}(f::F, grad!::G, xs::XT, storage::S)

Represents a composite function evaluated with stochastic gradient. f(θ, x) evaluates the loss for a single data point x and parameter θ. grad!(storage, θ, x) adds to storage the partial gradient with respect to data point x at parameter θ. xs must be an indexable iterable (Vector{Vector{Float64}} for instance). Functions using a StochasticObjective have optional keyword arguments rng, batch_size and full_evaluation controlling whether the function should be evaluated over all data points.

Note: grad! must not reset the storage to 0 before adding to it.

FrankWolfe.compute_gradient — Function

compute_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes the gradient of f at x. May return a reference to an internal storage.

FrankWolfe.compute_value — Function

compute_value(f::ObjectiveFunction, x; [kwargs...])

Computes the objective f at x.

FrankWolfe.compute_value_gradient — Method

compute_value_gradient(f::ObjectiveFunction, x; [kwargs...])

Computes in one call the pair (value, gradient) evaluated at x. By default, calls compute_value and compute_gradient with keywords kwargs passed down to both.

Custom vertex storage

FrankWolfe.RankOneMatrix — Type

RankOneMatrix{T, UT, VT}

Represents a rank-one matrix R = u * vt'. Composes like a charm.

FrankWolfe.ScaledHotVector — Type

ScaledHotVector{T}

Represents a vector of at most one value different from 0.

Utils

FrankWolfe.ConstantBatchIterator — Type

ConstantBatchIterator(batch_size)

Batch iterator always returning a constant batch size.

FrankWolfe.ConstantMomentumIterator — Type

ConstantMomentumIterator{T}

Iterator for momentum with a fixed damping value, always return the value and a dummy state.

FrankWolfe.ExpMomentumIterator — Type

ExpMomentumIterator{T}

Iterator for the momentum used in the variant of Stochastic Frank-Wolfe. Momentum coefficients are the values of the iterator: ρ_t = 1 - num / (offset + t)^exp

The state corresponds to the iteration count.

Source: Stochastic Conditional Gradient Methods: From Convex Minimization to Submodular Maximization Aryan Mokhtari, Hamed Hassani, Amin Karbasi, JMLR 2020.

FrankWolfe.IncrementBatchIterator — Type

IncrementBatchIterator(starting_batch_size, max_batch_size, [increment = 1])

Batch size starting at startingbatchsize and incrementing by increment at every iteration.

FrankWolfe.MonotonousNonConvexStepSize — Type

MonotonousNonConvexStepSize{F}

Represents a monotonous open-loop non-convex step size. Contains a halving factor N increased at each iteration until there is primal progress gamma = 1 / sqrt(t + 1) * 2^(-N)

FrankWolfe.MonotonousStepSize — Type

MonotonousStepSize{F}

Represents a monotonous open-loop step size. Contains a halving factor N increased at each iteration until there is primal progress gamma = 2 / (t + 2) * 2^(-N)

FrankWolfe._unsafe_equal — Method

_unsafe_equal(a, b)

Like isequal on arrays but without the checks. Assumes a and b have the same axes.

FrankWolfe.adaptive_step_size — Method

Slight modification of Adaptive Step Size strategy from https://arxiv.org/pdf/1806.05123.pdf

Note: direction is opposite to the improving direction norm(gradient, direction) > 0 TODO:

make emphasis aware and optimize

FrankWolfe.batchsize_iterate — Function

batchsize_iterate(iter::BatchSizeIterator) -> b

Method to implement for a batch size iterator of type BatchSizeIterator. Calling batchsize_iterate returns the next batch size and typically update the internal state of iter.

FrankWolfe.check_gradients — Function

Check if the gradient using finite differences matches the grad! provided.

FrankWolfe.line_search_wrapper — Method

line search wrapper NOTE: The stepsize is defined as x - gamma * d

Returns the step size gamma and the Lipschitz estimate L

FrankWolfe.momentum_iterate — Function

momentum_iterate(iter::MomentumIterator) -> ρ

Method to implement for a type MomentumIterator. Returns the next momentum value ρ and updates the iterator internal state.

FrankWolfe.plot_results — Method

plot_results

Given a series of list, generate subplots. listdatay -> contains a list of a list of lists (where each list refers to a subplot, and a list of lists refers to the y-values of the series inside a subplot). listdatax -> contains a list of a list of lists (where each list refers to a subplot, and a list of lists refers to the x-values of the series inside a subplot). So if we have one plot with two series, these might look like: listdatay = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]]] listdatax = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]]]

And if we have two plots, each with two series, these might look like: listdatay = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]], [[7, 8, 9, 10, 11, 12], [7, 8, 9, 10, 11, 12]]] listdatax = [[[1, 2, 3, 4, 5, 6], [1, 2, 3, 4, 5, 6]], [[7, 8, 9, 10, 11, 12], [7, 8, 9, 10, 11, 12]]]

listlabel -> contains the labels for the series that will be plotted, which has to have a length equal to the number of series that are being plotted: listlabel = ["Series 1", "Series 2"]

listaxisx -> contains the labels for the x-axis that will be plotted, which has to have a length equal to the number of subplots: listaxisx = ["x-axis plot 1", "x-axis plot 1"]

listaxisy -> Same as listaxisx but for the y-axis

xscalelog -> A list of values indicating the type of axes to use in each subplot, must be equal to the number of subplots: xscalelog = [:log, :identity]

yscalelog -> Same as xscalelog but for the y-axis

FrankWolfe.trajectory_callback — Method

trajectory_callback(storage)

Callback pushing the state at each iteration to the passed storage. The state data is only the 5 first fields, usually: (t,primal,dual,dual_gap,time)

Index

FrankWolfe.ActiveSet

FrankWolfe.ConstantBatchIterator

FrankWolfe.ConstantMomentumIterator

FrankWolfe.ExpMomentumIterator

FrankWolfe.IncrementBatchIterator

FrankWolfe.MonotonousNonConvexStepSize

FrankWolfe.MonotonousStepSize

FrankWolfe.ObjectiveFunction

FrankWolfe.RankOneMatrix

FrankWolfe.ScaledHotVector

FrankWolfe.SimpleFunctionObjective

FrankWolfe.StochasticObjective

Base.copy

FrankWolfe._unsafe_equal

FrankWolfe.active_set_argmin

FrankWolfe.active_set_argminmax

FrankWolfe.active_set_initialize!

FrankWolfe.active_set_update!

FrankWolfe.adaptive_step_size

FrankWolfe.batchsize_iterate

FrankWolfe.check_gradients

FrankWolfe.compute_active_set_iterate

FrankWolfe.compute_gradient

FrankWolfe.compute_value

FrankWolfe.compute_value_gradient

FrankWolfe.find_minmax_directions

FrankWolfe.line_search_wrapper

FrankWolfe.momentum_iterate

FrankWolfe.plot_results

FrankWolfe.trajectory_callback