Linear Minimization Oracles

Interface and wrappers

FrankWolfe.compute_extreme_point — Method

Compute the extreme point with a multi-vertex cache. store_cache indicates whether the newly-computed point should be stored in cache. greedy determines if we should return the first point with dot-product below threshold or look for the best one.

Norm balls

Simplex

Polytope

MathOptInterface

Index

FrankWolfe.compute_extreme_point